clean message
